// src/app/modules/client/enums/sse-events.enum.ts
export enum SseEvents {
  ACTUAL_SONG_SPTF = 'ACTUAL_SONG_SPTF',
  ACTUAL_VOTES_SPTF = 'ACTUAL_VOTES_SPTF',
  NEW_WINNER_SPTF = 'NEW_WINNER_SPTF',
  YOU_WIN_SONG_VOTE_SPTF = 'YOU_WIN_SONG_VOTE_SPTF',
  UPDATE_COINS = "UPDATE_COINS",
  NEW_PRIZE = "NEW_PRIZE",
  NEW_CLIENT_TO_PARTY = "NEW_CLIENT_TO_PARTY",
  LOGOUT_SPOTIFY = "LOGOUT_SPOTIFY",
}